Move GPUCard to Data Job Submission
